change addFloats return type

+++ src/AddNumbers.java
@@ -2,7 +2,7 @@ public class AddNumbers {
     public int addIntegers(int aNumber, int anotherNumber) {
         return aNumber + anotherNumber;
     }
-    public int addFloats(float aNumber, float anotherNumber) {
+    public float addFloats(float aNumber, float anotherNumber) {
         return aNumber + anotherNumber;
     }
 }
